Solved some warnings, continuing on start/stop/step of sandbox worlds
Continuing work on start/stop/step to work across all cases without crashes.
Solved some warnings which were good indications of possible bugs.
Fixed asset reference counts for shared instances from entity manager targets - now for each shared instance gets incremented with the number of entities that use it
Solved some problems with the recovery of the sandbox after module compilation. Needed to correctly reset the concurrency elements of the world and restore some system manager variables.

+++ ECSEngine/src/ECSEngine/Rendering/GraphicsHelpers.cpp
@@ -924,7 +924,7 @@ ECS_TEMPLATE_FUNCTION(Texture3D, function_name, Graphics*, Texture3D, bool); \
 		
 		ConstantBuffer vertex_constants = graphics->CreateConstantBuffer(sizeof(Matrix), true);
 		graphics->BindVertexConstantBuffer(vertex_constants);
-		GraphicsViewport cube_viewport = { 0.0f, 0.0f, face_size.x, face_size.y, 0.0f, 1.0f };
+		GraphicsViewport cube_viewport = { 0.0f, 0.0f, (float)face_size.x, (float)face_size.y, 0.0f, 1.0f };
 		graphics->BindViewport(cube_viewport);
 		graphics->DisableDepth();
 		graphics->DisableCulling();
